# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreatorOS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CreatorProfilePlatform =
  (typeof CreatorProfilePlatform)[keyof typeof CreatorProfilePlatform];

export const CreatorProfilePlatform = {
  instagram: "instagram",
  youtube: "youtube",
  x: "x",
} as const;

export interface CreatorProfile {
  id: number;
  name: string;
  handle: string;
  platform: CreatorProfilePlatform;
  followerCount: number;
  avatarEnabled: boolean;
  /** How the avatar greets and opens conversations */
  coreVibeConversational: string;
  /** The avatar's sales approach and product knowledge style */
  coreVibeOperational: string;
  /** How the avatar builds trust and handles objections */
  coreVibeRelational: string;
  /** Boundaries and values the avatar upholds */
  coreVibeEthos: string;
  /** Comma-separated Hinglish slang words the avatar uses */
  signatureSlang: string;
  profileImage?: string | null;
  createdAt: string;
}

export type UpdateCreatorProfileBodyPlatform =
  (typeof UpdateCreatorProfileBodyPlatform)[keyof typeof UpdateCreatorProfileBodyPlatform];

export const UpdateCreatorProfileBodyPlatform = {
  instagram: "instagram",
  youtube: "youtube",
  x: "x",
} as const;

export interface UpdateCreatorProfileBody {
  name?: string;
  handle?: string;
  platform?: UpdateCreatorProfileBodyPlatform;
  avatarEnabled?: boolean;
  coreVibeConversational?: string;
  coreVibeOperational?: string;
  coreVibeRelational?: string;
  coreVibeEthos?: string;
  signatureSlang?: string;
  profileImage?: string | null;
}

export type CampaignStatus =
  (typeof CampaignStatus)[keyof typeof CampaignStatus];

export const CampaignStatus = {
  active: "active",
  paused: "paused",
  completed: "completed",
  draft: "draft",
} as const;

export type CampaignPlatform =
  (typeof CampaignPlatform)[keyof typeof CampaignPlatform];

export const CampaignPlatform = {
  instagram: "instagram",
  youtube: "youtube",
  x: "x",
} as const;

export interface Campaign {
  id: number;
  brandName: string;
  productName: string;
  status: CampaignStatus;
  platform: CampaignPlatform;
  /** Total brand deal value in INR */
  dealValue: number;
  startDate: string;
  endDate?: string | null;
  totalClicks: number;
  totalConversions: number;
  totalRevenue: number;
  conversionRate: number;
  createdAt: string;
}

export type CreateCampaignBodyPlatform =
  (typeof CreateCampaignBodyPlatform)[keyof typeof CreateCampaignBodyPlatform];

export const CreateCampaignBodyPlatform = {
  instagram: "instagram",
  youtube: "youtube",
  x: "x",
} as const;

export interface CreateCampaignBody {
  brandName: string;
  productName: string;
  platform: CreateCampaignBodyPlatform;
  dealValue: number;
  startDate: string;
  endDate?: string | null;
}

export type UpdateCampaignBodyStatus =
  (typeof UpdateCampaignBodyStatus)[keyof typeof UpdateCampaignBodyStatus];

export const UpdateCampaignBodyStatus = {
  active: "active",
  paused: "paused",
  completed: "completed",
  draft: "draft",
} as const;

export interface UpdateCampaignBody {
  brandName?: string;
  productName?: string;
  status?: UpdateCampaignBodyStatus;
  dealValue?: number;
  endDate?: string | null;
}

export interface Product {
  id: number;
  name: string;
  description?: string | null;
  price: number;
  category: string;
  imageUrl?: string | null;
  inStock: boolean;
  totalSales: number;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  description?: string | null;
  price: number;
  category: string;
  imageUrl?: string | null;
  inStock?: boolean;
}

export interface UpdateProductBody {
  name?: string;
  description?: string | null;
  price?: number;
  category?: string;
  inStock?: boolean;
  imageUrl?: string | null;
}

export interface TrackingLink {
  id: number;
  shortCode: string;
  destinationUrl: string;
  campaignId?: number | null;
  productId?: number | null;
  label?: string | null;
  totalClicks: number;
  uniqueClicks: number;
  conversions: number;
  revenue: number;
  createdAt: string;
}

export interface CreateLinkBody {
  destinationUrl: string;
  campaignId?: number | null;
  productId?: number | null;
  label?: string | null;
}

export type ConversationPlatform =
  (typeof ConversationPlatform)[keyof typeof ConversationPlatform];

export const ConversationPlatform = {
  instagram: "instagram",
  youtube: "youtube",
  x: "x",
} as const;

export type ConversationStatus =
  (typeof ConversationStatus)[keyof typeof ConversationStatus];

export const ConversationStatus = {
  active: "active",
  converted: "converted",
  dropped: "dropped",
  pending: "pending",
} as const;

export interface Conversation {
  id: number;
  followerHandle: string;
  platform: ConversationPlatform;
  status: ConversationStatus;
  triggerComment: string;
  messageCount: number;
  linkSent: boolean;
  converted: boolean;
  revenueGenerated: number;
  startedAt: string;
  lastMessageAt: string;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  avatar: "avatar",
  follower: "follower",
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  sentAt: string;
}

export interface ConversationDetail {
  conversation: Conversation;
  messages: Message[];
}

export interface DashboardSummary {
  totalRevenue: number;
  totalConversions: number;
  totalDmsHandled: number;
  avgConversionRate: number;
  activeCampaigns: number;
  totalBrands: number;
  revenueThisMonth: number;
  conversionsThisMonth: number;
  /** Month-over-month revenue growth percentage */
  revenueGrowthPct: number;
}

export interface FunnelStat {
  campaignId: number;
  brandName: string;
  comments: number;
  dmsInitiated: number;
  linksClicked: number;
  purchases: number;
  conversionRate: number;
}

export interface RevenueTrendPoint {
  date: string;
  revenue: number;
  conversions: number;
}

export interface TopProduct {
  productId: number;
  name: string;
  category: string;
  totalSales: number;
  totalRevenue: number;
  imageUrl?: string | null;
}