# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CreatorOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /creator/profile:
    get:
      operationId: getCreatorProfile
      tags: [creator]
      summary: Get creator profile and avatar config
      responses:
        "200":
          description: Creator profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorProfile"
    put:
      operationId: updateCreatorProfile
      tags: [creator]
      summary: Update creator profile and avatar vibe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCreatorProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorProfile"
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all brand deal campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    CreatorProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        handle:
          type: string
        platform:
          type: string
          enum: [instagram, youtube, x]
        followerCount:
          type: integer
        avatarEnabled:
          type: boolean
        coreVibeConversational:
          type: string
          description: How the avatar greets and opens conversations
        coreVibeOperational:
          type: string
          description: The avatar's sales approach and product knowledge style
        coreVibeRelational:
          type: string
          description: How the avatar builds trust and handles objections
        coreVibeEthos:
          type: string
          description: Boundaries and values the avatar upholds
        signatureSlang:
          type: string
          description: Comma-separated Hinglish slang words the avatar uses
        profileImage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - handle
        - platform
        - followerCount
        - avatarEnabled
        - coreVibeConversational
        - coreVibeOperational
        - coreVibeRelational
        - coreVibeEthos
        - signatureSlang
        - createdAt
    UpdateCreatorProfileBody:
      type: object
      properties:
        name:
          type: string
        handle:
          type: string
        platform:
          type: string
          enum: [instagram, youtube, x]
        avatarEnabled:
          type: boolean
        coreVibeConversational:
          type: string
        coreVibeOperational:
          type: string
        coreVibeRelational:
          type: string
        coreVibeEthos:
          type: string
        signatureSlang:
          type: string
        profileImage:
          type: string
          nullable: true
    Campaign:
      type: object
      properties:
        id:
          type: integer
        brandName:
          type: string
        productName:
          type: string
        status:
          type: string
          enum: [active, paused, completed, draft]
        platform:
          type: string
          enum: [instagram, youtube, x]
        dealValue:
          type: number
          description: Total brand deal value in INR
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        totalClicks:
          type: integer
        totalConversions:
          type: integer
        totalRevenue:
          type: number
        conversionRate:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - brandName
        - productName
        - status
        - platform
        - dealValue
        - startDate
        - totalClicks
        - totalConversions
        - totalRevenue
        - conversionRate
        - createdAt
    CreateCampaignBody:
      type: object
      required:
        - brandName
        - productName
        - platform
        - dealValue
        - startDate
      properties:
        brandName:
          type: string
        productName:
          type: string
        platform:
          type: string
          enum: [instagram, youtube, x]
        dealValue:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
    UpdateCampaignBody:
      type: object
      properties:
        brandName:
          type: string
        productName:
          type: string
        status:
          type: string
          enum: [active, paused, completed, draft]
        dealValue:
          type: number
        endDate:
          type: string
          format: date
          nullable: true
...