# lib/api-zod/src/generated/types/updateCampaignBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreatorOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCampaignBodyStatus } from "./updateCampaignBodyStatus";

export interface UpdateCampaignBody {
  brandName?: string;
  productName?: string;
  status?: UpdateCampaignBodyStatus;
  dealValue?: number;
  endDate?: Date | null;
}