# lib/api-zod/src/generated/types/updateCreatorProfileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CreatorOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCreatorProfileBodyPlatform } from "./updateCreatorProfileBodyPlatform";

export interface UpdateCreatorProfileBody {
  name?: string;
  handle?: string;
  platform?: UpdateCreatorProfileBodyPlatform;
  avatarEnabled?: boolean;
  coreVibeConversational?: string;
  coreVibeOperational?: string;
  coreVibeRelational?: string;
  coreVibeEthos?: string;
  signatureSlang?: string;
  profileImage?: string | null;
}